backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.16
langchain-openai==0.1.3
openai==1.23.2
